pyproject.toml
[project]
name = "hexin-pixel-font"
version = "0.1.0"
description = ""
authors = [
    {name = "diaowinner",email = "diaowinner@qq.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich (>=14.1.0,<15.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
    "pixel-font-builder (>=0.0.37,<0.0.38)"
]

[tool.poetry]
packages = [{include = "hexin_pixel_font", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
